arch/sim: Rename sim_video to sim_camera
