+++ arch/sim/src/sim/posix/sim_host_v4l2.c
@@ -208,7 +208,7 @@ int host_video_start_capture(struct host_video_dev_s *vdev)
         }
 
       vdev->addrs[i] = mmap(NULL, buf.length, PROT_READ | PROT_WRITE,
-                           MAP_SHARED, vdev->fd, buf.m.offset);
+                            MAP_SHARED, vdev->fd, buf.m.offset);
       if (vdev->addrs[i] == MAP_FAILED)
         {
           perror("Mmap failed");
